MappingBuilder : renommage et réécriture suite à interface
- Une interface MappingBuilder (dans core) a été créée à partir du code
qu'on avait
- Ré-écriture du mapping builder elasticsearch en conséquence
(surppression de la phpdoc redondante) et renommage
- Repasse ignore_malformed à true : si on a la moindre date (ou nombre)
invalide, ça rejette tout l'enregistrement, ce qui n'est pas acceptable.
- Ajout de la méthode reset()
